Move request and response validation to appropriate files
